Structure for code refactor
